add local column type to config

+++ benchmark/src/test/java/com/hufudb/openhufu/benchmark/OpenHuFuBenchmarkTest.java
@@ -133,7 +133,7 @@ public class OpenHuFuBenchmarkTest {
     assertEquals(90, count);
     dataset.close();
   }
-@Test
+  @Test
   public void testLeftJoin() {
     String leftTableName = TPCHTableName.SUPPLIER.getName();
     LeafPlan leftPlan = new LeafPlan();
@@ -199,7 +199,7 @@ public class OpenHuFuBenchmarkTest {
     assertEquals(138, count);
     dataset.close();
   }
-@Test
+  @Test
   public void testOuterJoin() {
     String leftTableName = TPCHTableName.SUPPLIER.getName();
     LeafPlan leftPlan = new LeafPlan();
@@ -232,6 +232,7 @@ public class OpenHuFuBenchmarkTest {
     assertEquals(138, count);
     dataset.close();
   }
+
   @Test
   public void testSemiJoin() {
     //todo
@@ -317,11 +318,12 @@ public class OpenHuFuBenchmarkTest {
 
   @Test
   public void testMax() {
+
     String tableName = TPCHTableName.SUPPLIER.getName();
     LeafPlan plan = new LeafPlan();
     plan.setTableName(tableName);
     plan.setAggExps(ImmutableList.of(
-            ExpressionFactory.createAggFunc(ColumnType.DOUBLE, Modifier.PUBLIC, AggFuncType.MAX.getId(),
+            ExpressionFactory.createAggFunc(ColumnType.LONG, Modifier.PUBLIC, AggFuncType.MAX.getId(),
                     ImmutableList.of(ExpressionFactory.createInputRef(0, ColumnType.LONG, Modifier.PUBLIC)))));
 
     DataSet dataset = user.executeQuery(plan);
